ESC-710 factored out confirm email post handling
